refactor search page into index page
